麻酔ポーション更新 メッセージテキストの表示追加
