+++ Assets/Scenes/GameScene.unity
@@ -4139,6 +4139,7 @@ MonoBehaviour:
   - {fileID: 11400000, guid: 70c7492d010a7fb4d9501792b1fda355, type: 2}
   - {fileID: 11400000, guid: ab35e5af54dbebf4daec2b12693346d8, type: 2}
   - {fileID: 11400000, guid: c67afd4b9bf232646a5f63c4aa7aceb6, type: 2}
+  - {fileID: 11400000, guid: 9eebfc4e3c71b904c800c45ae043383e, type: 2}
 --- !u!4 &959393679
 Transform:
   m_ObjectHideFlags: 0
@@ -5515,7 +5516,7 @@ RectTransform:
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
   m_AnchorMin: {x: 0, y: 1}
   m_AnchorMax: {x: 1, y: 1}
-  m_AnchoredPosition: {x: 0, y: 200.00053}
+  m_AnchoredPosition: {x: 0, y: 200.00012}
   m_SizeDelta: {x: 0, y: 800}
   m_Pivot: {x: 0, y: 1}
 --- !u!1 &1324895147
@@ -8632,8 +8633,8 @@ MonoBehaviour:
   m_TargetGraphic: {fileID: 1804400923}
   m_HandleRect: {fileID: 1804400922}
   m_Direction: 2
-  m_Value: -0.0000030517588
-  m_Size: 0.7499196
+  m_Value: -0.00000061035155
+  m_Size: 0.62166506
   m_NumberOfSteps: 0
   m_OnValueChanged:
     m_PersistentCalls:
